+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.3] - 2025-05-13
+
+### Fixed
+
+- Automatically close connection for text over TCP entity after the command has been send
+
+### Changed
+
+- Updated pyinstaller build image to r2-pyinstaller:3.11.12-0.3.0
+- Updated dependencies
+
 ## [0.7.2] - 2025-04-04
 
 ### Fixed